requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
google-genai
python-dotenv
streamlit
pandas